Extended Graph

MeGraPDE.MetricGraphs.discretize_function — Function

discretize_function(Γ::MetricGraph, u::Vector{Function}, h_max::Number)

Return discretized version of 'u' on the extended graph of 'Γ' with step size 'h_max' on the edges.

discretize_function(Γ::MetricGraph, u::Vector{Function}, Nx_vec::Vector)

Return discretized version of 'u' on the extended graph of 'Γ' with inner grid points in 'Nx_vec'.

discretize_function(Γ::EquilateralMetricGraph, u::Vector{Function}, h_max::Number)

Equilateral version.

MeGraPDE.MetricGraphs.extended_incidence_matrix — Function

extended_incidence_matrix(Γ::MetricGraph, h_max::Number)

Return extended incidence matrix of 'Γ' with maximal step length 'h_max' per edge.

Construction of incidence matrix via kron-products according to (AB) (section 4.1), see also (W), section 3.1 for a summary.

extended_incidence_matrix(Γ::EquilateralMetricGraph, h_max::Number)

Equilateral version with some simplifications.

MeGraPDE.MetricGraphs.extended_laplacian — Function

extended_laplacian(Γ::EquilateralMetricGraph, k::Int)

Compute extended graph Laplacian matrix of 'Γ' with 'k' artificial vertices on each edge.

k inner vertices means each edge is partitioned in k+1 subdivision. The construction of the graph Laplacian relies on the same manipulations of the original graph as in the extendedincidencematrix routine. Here, however, the Laplacian matrix L=NN^T is returned and some simplifications due to the equilateral edge length apply.