Test Problems

Different type of test problems of abstract type TestProblem can be constructed.

The following types are available:

EllipticTestProblem

TPGeneralizedHeat

TPReactionDiffusion

Depending on the specific problem, a right-hand side, reaction-term and/or initial conditions must be set. If available, the exact solution and its derivative can be indicated. Otherwise, 'Nothing' must be specified.

Elliptic Test Problems

The elliptic equation

\[\mathcal{H}( u(x)) + \nu u(x) = f(x)\]

is represented by the TestProblem type

MeGraPDE.EllipticTestProblem — Type

Elliptic Test Problem

pot

Potential

Γ

Metric Graph

rhs

Right-hand side

u

Exact solution

u_deriv

Derivative of exact solution

The following exemplary test problems are predefined:

MeGraPDE.TestProblem242 — Constant

TestProblem242

Elliptic test problem on 5-star graph with equilateral edge length π+π/2.

MeGraPDE.TestProblem243 — Constant

TestProblem243

Elliptic test problem diamond graph with equilateral edge length 2π.

MeGraPDE.TestProblem711 — Constant

TestProblem711

\[\mathcal{H} u + u = f\]

on star graph with

\[f = -\frac{(\exp(-(x-x_0)^2/s^2) \cdot (4 (x-x_0)^2-2s^2)}{s^4} + \exp \left(- \frac{(x-x_0)^2}{s^2} \right)\]

on one edge and exact solution

\[\exp \left(- \frac{(x-x_0)^2}{s^2} \right).\]

Parabolic Test Problems

Generalized Heat Equation

The generalized heat equation

\[\frac{\partial u}{\partial t}(x,t) + \mathcal{H}( u(x,t) = f(x,t))\]

on $\Gamma$ subject to Neumann-Kirchhoff conditions and initial condition $u^0$ is represented by the TestProblem type

MeGraPDE.TPGeneralizedHeat — Type

Test Problem related to Generalized Heat Equation on Γ.

Γ

Metric Graph

u0

Initial Condition

rhs

right-hand side

u

Exact solution

u_deriv

Derivative of exact solution

The following exemplary test problems are predefined:

MeGraPDE.TestProblem244 — Constant

Heat equation on star graph with eigenfunction $\phi_5$ as initial condition.

MeGraPDE.TestProblem245 — Constant

Heat equation on diamond graph with eigenfunction $\phi_3$ as initial condition.

MeGraPDE.TestProblem721 — Constant

Heat equation with Gaussian-type inital condition.