Spectral Galerkin Method

MeGraPDE.projection_coefficients — Function

projection_coefficients(Γ::MetricGraph, σ::QuantumGraphEigen, u::Vector{Function})

Compute projection coefficients 'coefs' of orthogonal projection of 'u' with standard QuadGK quadrature.

projection_coefficients(Γ::EquilateralMetricGraph, σ::QuantumGraphEigen, u::Vector{Function})

Equilateral version.

MeGraPDE.projection_coefficients_filon — Function

projection_coefficients_filon(Γ::EquilateralMetricGraph, σ::QuantumGraphEigen, u::Vector{Function}, quad_nodes::Int)

Compute projection coefficients of 'u' on 'Γ' for eigenfunction in 'σ' with 'quad_nodes' quadrature nodes.

Uses (matrix-free) Filon-Quadrature as described in [W], Section 4.3.2.

MeGraPDE.spectral_solution — Function

spectral_solution(Γ::AbstractMetricGraph, σ::QuantumGraphEigen, coef::Vector)

Explicitly construct spectral solution u_Q on 'Γ' from eigenbasis 'σ' and coefficents 'coef'.

MeGraPDE.spectral_solver — Function

spectral_solver(TP::EllipticTestProblem, K::Int; filon=false)

Compute coefficents of the spectral Galerkin solution of 'TP' with eigenfunction basis of size 'K'.

If filon=true, the more economic filon-quadrature is applied instead of QuadGK.

spectral_solver(TP::TPGeneralizedHeat, T::Number, K::Int; filon=false)

Compute coefficents of the spectral Galerkin solution of 'TP' at time 'T' with eigenfunction basis of size 'K'.

If filon=true, the more economic filon-quadrature is applied instead of QuadGK.