Non-Equilateral Graphs

Full Doc

MeGraPDE.QuantumGraphSpectra.equilateral_floor_approximation — Function

equilateral_floor_approximation(Γ::MetricGraph, h::Number)

Compute equilateral floor approximation of 'Γ' with equilateral edge length 'h'

MeGraPDE.QuantumGraphSpectra.equilateral_ceil_approximation — Function

equilateral_ceil_approximation(Γ::MetricGraph, h::Number)

Compute equilateral ceil approximation of 'Γ' with equilateral edge length 'h'

MeGraPDE.QuantumGraphSpectra.eigvals_equilateral_representation — Function

eigvals_equilateral_representation(Γ::MetricGraph, h::Number)

Compute the exact eigenvalues of 'Γ' by an equilateral representation with edge length 'h'

MeGraPDE.QuantumGraphSpectra.approx_lowest_level — Function

approx_lowest_level(Γ::MetricGraph, h_min::Number; Q=2)

Compute eigenvalue approximations by equilateral ceil and floor approximations of the first 'Q' eigenvalues at the lowest discretization level 'h_min' in the nested iteration.

MeGraPDE.QuantumGraphSpectra.nested_iteration_eigenvalue_approximation — Function

nested_iteration_eigenvalue_approximation(Γ::MetricGraph; lev_zero=0, lev_max=7, Q=2, save_each_lev=false)

Approximate first 'Q' eigenvalues of 'Γ' via equilateral approximations using a nested itertation approach.

MeGraPDE.QuantumGraphSpectra.H_matrix — Function

H_matrix(z::Number, Γ::MetricGraph)

Compute H(z) for a metric graph 'Γ'.

H_matrix(z::Number, bfN::SparseMatrixCSC, ℓ_vec::Vector)

Compute H(z) for a graph with incidence matrix 'Inc' and edge length 'ℓ_vec'.

MeGraPDE.QuantumGraphSpectra.newton_trace — Function

newton_trace(Γ::MetricGraph, z_start::Number)

Newton-trace iteration to determine roots of det(H(z)).

MeGraPDE.QuantumGraphSpectra.nested_iteration_newton_trace — Function

nested_iteration_newton_trace(Γ::MetricGraph; lev_zero=0, lev_max=7, Q=5, save_each_lev=false, return_eigvecs=false)

Conduct nested iteration newton trace algorithm to find the first 'Q' eigenvalues of 'Γ'.